build_assets/aws_lambda_dockerfiles/requirements.txt
#Python 3.10.11
tensorflow-cpu==2.11.1
protobuf==3.19.6
boto3>=1.26.134